digital-asset-protection/requirements.txt
librosa
soundfile
pyacoustid          # Chromaprint bindings
transformers
torch
open-clip-torch
fastapi
uvicorn
requests
yt-dlp
python-dotenv
numpy
pillow
python-multipart